updated python tests
