+++ .github/workflows/python-app.yml
@@ -56,9 +56,6 @@ jobs:
 
       - name: List coverage directory contents
         run: ls -la coverage
-
-      - name: View Jest JUnit report
-        run: cat coverage/sonar-report.xml
   
       - name: SonarCloud Scan
         uses: SonarSource/sonarcloud-github-action@master
