+++ openwaves/tests/test_sessions.py
@@ -5,9 +5,8 @@
 
 from datetime import datetime
 from openwaves import db
-from openwaves.imports import User, Pool, ExamSession, ExamRegistration
+from openwaves.imports import User, Pool, ExamSession
 from openwaves.tests.test_auth import login
-
 
 #############################
 #                           #
@@ -32,7 +31,6 @@ def test_sessions_page_access(client, ve_user):
     assert response.status_code == 200
     assert b'Test Sessions' in response.data
 
-
 def test_sessions_page_access_as_regular_user(client):
     """Test ID: UT-62
     Negative test: Ensures that a regular user cannot access the sessions page.
@@ -47,7 +45,6 @@ def test_sessions_page_access_as_regular_user(client):
     response = client.get('/ve/sessions', follow_redirects=True)
     assert b'Access denied.' in response.data
 
-
 def test_sessions_page_access_not_logged_in(client):
     """Test ID: UT-63
     Negative test: Ensures that unauthenticated users cannot access the sessions page.
@@ -61,7 +58,6 @@ def test_sessions_page_access_not_logged_in(client):
     """
     response = client.get('/ve/sessions', follow_redirects=True)
     assert b'Please log in to access this page.' in response.data
-
 
 def test_create_session_success(client, ve_user):
     """Test ID: UT-64
@@ -132,7 +128,6 @@ def test_create_session_success(client, ve_user):
         assert session.gen_pool_id == general_pool.id
         assert session.extra_pool_id == extra_pool.id
 
-
 def test_create_session_missing_fields(client, ve_user):
     """Test ID: UT-65
     Negative test: Verifies that creating a session with missing fields returns an error.
@@ -155,7 +150,6 @@ def test_create_session_missing_fields(client, ve_user):
     assert response.status_code == 400
     assert response.is_json
     assert 'All fields are required.' in response.get_json()['error']
-
 
 def test_create_session_access_as_regular_user(client):
     """Test ID: UT-66
@@ -180,7 +174,6 @@ def test_create_session_access_as_regular_user(client):
     assert response.status_code == 302
     assert '/auth/logout' in response.headers['Location']
 
-
 def test_create_session_not_logged_in(client):
     """Test ID: UT-67
     Negative test: Ensures that unauthenticated users cannot create a test session.
@@ -199,7 +192,6 @@ def test_create_session_not_logged_in(client):
         'extra_pool': '1'
     }, follow_redirects=True)
     assert b'Please log in to access this page.' in response.data
-
 
 def test_open_session_success(client, ve_user):
     """Test ID: UT-68
@@ -272,7 +264,6 @@ def test_open_session_success(client, ve_user):
         assert session.status is True
         assert session.start_time is not None
 
-
 def test_open_session_not_found(client, ve_user):
     """Test ID: UT-69
     Negative test: Ensures that trying to open a non-existent session returns a 404 error.
@@ -295,7 +286,6 @@ def test_open_session_not_found(client, ve_user):
     assert response.is_json
     json_data = response.get_json()
     assert json_data['error'] == "Session not found."
-
 
 def test_close_session_success(client, ve_user):
     """Test ID: UT-70
@@ -377,7 +367,6 @@ def test_close_session_success(client, ve_user):
         assert closed_session.status is False, "The session is still open."
         assert closed_session.end_time is not None, "The session end_time was not set."
 
-
 def test_close_session_not_found(client, ve_user):
     """Test ID: UT-71
     Negative test: Ensures that trying to close a non-existent session returns a 404 error.
@@ -401,7 +390,6 @@ def test_close_session_not_found(client, ve_user):
     json_data = response.get_json()
     assert json_data['error'] == "Session not found."
 
-
 def test_open_session_access_as_regular_user(client):
     """Test ID: UT-72
     Negative test: Ensures that a regular user cannot open a test session.
@@ -415,7 +403,6 @@ def test_open_session_access_as_regular_user(client):
     login(client, 'TESTUSER', 'testpassword')
     response = client.post('/ve/open_session/1', follow_redirects=True)
     assert b'Access denied.' in response.data
-
 
 def test_close_session_access_as_regular_user(client):
     """Test ID: UT-73
@@ -431,7 +418,6 @@ def test_close_session_access_as_regular_user(client):
     response = client.post('/ve/close_session/1', follow_redirects=True)
     assert b'Access denied.' in response.data
 
-
 def test_open_session_not_logged_in(client):
     """Test ID: UT-74
     Negative test: Ensures that an unauthenticated user cannot open a test session.
@@ -445,7 +431,6 @@ def test_open_session_not_logged_in(client):
     response = client.post('/ve/open_session/1', follow_redirects=True)
     assert b'Please log in to access this page.' in response.data
 
-
 def test_close_session_not_logged_in(client):
     """Test ID: UT-74
     Negative test: Ensures that an unauthenticated user cannot open a test session.
@@ -458,7 +443,6 @@ def test_close_session_not_logged_in(client):
     """
     response = client.post('/ve/close_session/1', follow_redirects=True)
     assert b'Please log in to access this page.' in response.data
-
 
 #############################
 #                           #
@@ -550,333 +534,3 @@ def test_sessions_route_as_unauthorized_user(client):
     assert response.status_code == 200
     print(response.data)
     assert b'Please log in to access this page.' in response.data
-
-def test_register_route_success(client, app):
-    """Test ID: UT-93
-    Test successful registration for an exam session.
-
-    This test ensures that a user with role 1 can successfully register for an exam element.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The registration is created in the database.
-        - The user receives a success message.
-    """
-    with app.app_context():
-        # Log in as the test user (role 1)
-        user = User.query.filter_by(username="TESTUSER").first()
-        login(client, user.username, "testpassword")
-
-        # Create an exam session
-        pool = Pool(
-            name="Test Pool",
-            element=2,
-            start_date=datetime(2023, 1, 1),
-            end_date=datetime(2026, 12, 31)
-        )
-        db.session.add(pool)
-        db.session.commit()
-
-        exam_session = ExamSession(
-            session_date=datetime.now(),
-            tech_pool_id=pool.id,
-            gen_pool_id=pool.id,
-            extra_pool_id=pool.id,
-            status=True
-        )
-        db.session.add(exam_session)
-        db.session.commit()
-
-        # Submit a registration request
-        response = client.post('/register', data={
-            'session_id': exam_session.id,
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'Successfully registered for the Tech exam.' in response.data
-
-        # Verify registration in the database
-        registration = ExamRegistration.query.filter_by(
-            session_id=exam_session.id,
-            user_id=user.id
-        ).first()
-        assert registration is not None
-        assert registration.tech is True
-
-def test_register_route_already_registered(client, app):
-    """Test ID: UT-94
-    Test registration when the user is already registered for the exam element.
-
-    This test ensures that attempting to register again for the same exam element 
-    results in an appropriate error message.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The user receives an error message.
-        - No duplicate registrations are created.
-    """
-    with app.app_context():
-        # Log in as the test user (role 1)
-        user = User.query.filter_by(username="TESTUSER").first()
-        login(client, user.username, "testpassword")
-
-        # Create an exam session and register the user
-        pool = Pool(
-            name="Test Pool",
-            element=2,
-            start_date=datetime(2023, 1, 1),
-            end_date=datetime(2026, 12, 31)
-        )
-        db.session.add(pool)
-        db.session.commit()
-
-        exam_session = ExamSession(
-            session_date=datetime.now(),
-            tech_pool_id=pool.id,
-            gen_pool_id=pool.id,
-            extra_pool_id=pool.id,
-            status=True
-        )
-        db.session.add(exam_session)
-        db.session.commit()
-
-        # Initial registration
-        registration = ExamRegistration(
-            session_id=exam_session.id,
-            user_id=user.id,
-            tech=True
-        )
-        db.session.add(registration)
-        db.session.commit()
-
-        # Attempt to register again for the same exam element
-        response = client.post('/register', data={
-            'session_id': exam_session.id,
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'You are already registered for the Tech exam.' in response.data
-
-def test_register_route_missing_data(client, app):
-    """Test ID: UT-95
-    Test registration with missing form data.
-
-    This test ensures that attempting to register without providing required form data 
-    results in an appropriate error message.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The user receives an error message about missing information.
-    """
-    with app.app_context():
-        # Log in as the test user (role 1)
-        user = User.query.filter_by(username="TESTUSER").first()
-        login(client, user.username, "testpassword")
-
-        # Attempt to register without session_id
-        response = client.post('/register', data={
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'Invalid registration request. Missing required information.' in response.data
-
-def test_cancel_registration_success(client, app):
-    """Test ID: UT-96
-    Test successful cancellation of an exam registration.
-
-    This test ensures that a user with role 1 can successfully cancel their registration 
-    for an exam element.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The registration is updated in the database.
-        - The user receives a success message.
-    """
-    with app.app_context():
-        # Log in as the test user (role 1)
-        user = User.query.filter_by(username="TESTUSER").first()
-        login(client, user.username, "testpassword")
-
-        # Create an exam session and register the user
-        pool = Pool(
-            name="Test Pool",
-            element=2,
-            start_date=datetime(2023, 1, 1),
-            end_date=datetime(2026, 12, 31)
-        )
-        db.session.add(pool)
-        db.session.commit()
-
-        exam_session = ExamSession(
-            session_date=datetime.now(),
-            tech_pool_id=pool.id,
-            gen_pool_id=pool.id,
-            extra_pool_id=pool.id,
-            status=True
-        )
-        db.session.add(exam_session)
-        db.session.commit()
-
-        # Initial registration
-        registration = ExamRegistration(
-            session_id=exam_session.id,
-            user_id=user.id,
-            tech=True
-        )
-        db.session.add(registration)
-        db.session.commit()
-
-        # Submit a cancellation request
-        response = client.post('/cancel_registration', data={
-            'session_id': exam_session.id,
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'Successfully canceled registration for the Tech exam.' in response.data
-
-        # Verify registration in the database
-        updated_registration = ExamRegistration.query.filter_by(
-            session_id=exam_session.id,
-            user_id=user.id
-        ).first()
-        assert updated_registration.tech is False
-
-def test_cancel_registration_not_registered(client, app):
-    """Test ID: UT-97
-    Test cancellation when the user is not registered for the exam element.
-
-    This test ensures that attempting to cancel a registration for an exam element 
-    the user is not registered for results in an appropriate error message.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The user receives an error message.
-    """
-    with app.app_context():
-        # Log in as the test user (role 1)
-        user = User.query.filter_by(username="TESTUSER").first()
-        login(client, user.username, "testpassword")
-
-        # Create an exam session
-        pool = Pool(
-            name="Test Pool",
-            element=2,
-            start_date=datetime(2023, 1, 1),
-            end_date=datetime(2026, 12, 31)
-        )
-        db.session.add(pool)
-        db.session.commit()
-
-        exam_session = ExamSession(
-            session_date=datetime.now(),
-            tech_pool_id=pool.id,
-            gen_pool_id=pool.id,
-            extra_pool_id=pool.id,
-            status=True
-        )
-        db.session.add(exam_session)
-        db.session.commit()
-
-        # Attempt to cancel registration without being registered
-        response = client.post('/cancel_registration', data={
-            'session_id': exam_session.id,
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'You are not registered for the Tech exam.' in response.data
-
-def test_register_route_invalid_role(client, app):
-    """Test ID: UT-98
-    Test registration with a user who does not have role 1.
-
-    This test ensures that a user with a role other than 1 cannot register for an exam.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The user is redirected with an access denied message.
-    """
-    with app.app_context():
-        # Create and log in as a user with role 2 (VE)
-        ve_user = User(
-            username="VEUSER2",
-            first_name="VE",
-            last_name="User",
-            email="veuser2@example.com",
-            password="vepassword",
-            role=2,
-            active=True
-        )
-        db.session.add(ve_user)
-        db.session.commit()
-
-        login(client, ve_user.username, "vepassword")
-
-        # Attempt to register
-        response = client.post('/register', data={
-            'session_id': '1',
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'Please log in to access this page.' in response.data
-
-def test_cancel_registration_invalid_role(client, app):
-    """Test ID: UT-99
-    Test cancellation with a user who does not have role 1.
-
-    This test ensures that a user with a role other than 1 cannot cancel an exam registration.
-
-    Args:
-        client: The test client instance.
-        app: The Flask application instance.
-
-    Asserts:
-        - The user is redirected with an access denied message.
-    """
-    with app.app_context():
-        # Create and log in as a user with role 2 (VE)
-        ve_user = User(
-            username="VEUSER3",
-            first_name="VE",
-            last_name="User",
-            email="veuser3@example.com",
-            password="vepassword",
-            role=2,
-            active=True
-        )
-        db.session.add(ve_user)
-        db.session.commit()
-
-        login(client, ve_user.username, "vepassword")
-
-        # Attempt to cancel registration
-        response = client.post('/cancel_registration', data={
-            'session_id': '1',
-            'exam_element': '2'
-        }, follow_redirects=True)
-
-        assert response.status_code == 200
-        assert b'Please log in to access this page.' in response.data
